Don't auto update Kotlin version + KSP
This adds component selection rules, so that only
the version specified in the toml file is
accepted for the Kotlin version, it's never
automatically updated as a result.

As KSP requires a specific Kotlin version, any
version that does not match the kotlin version
is also rejected.

To update both we can manually bump the Kotlin
version in the toml file.